Total Menu Fix
Changed the menuSystem stack from an object to a pointer and passed that as a parameter instead of passing it as a value which was the reason for using the dummy menu objects and doubling up on pushes. Not sure why I originally decided to pass them as an object instead of a pointer.